+++ packages/fhir-4/src/profiles/ResearchElementDefinition.ts
@@ -15,9 +15,9 @@ export type ResearchElementDefinition_Props = {
     comment?: string[];
     contact?: FHIR.ContactDetail[];
     contained?: any[];
-    copyright?: FHIR.markdown;
+    copyright?: string;
     date?: string;
-    description?: FHIR.markdown;
+    description?: string;
     editor?: FHIR.ContactDetail[];
     effectivePeriod?: FHIR.Period;
     endorser?: FHIR.ContactDetail[];
@@ -34,7 +34,7 @@ export type ResearchElementDefinition_Props = {
     modifierExtension?: FHIR.Extension[];
     name?: string;
     publisher?: string;
-    purpose?: FHIR.markdown;
+    purpose?: string;
     relatedArtifact?: FHIR.RelatedArtifact[];
     reviewer?: FHIR.ContactDetail[];
     shortTitle?: string;