+++ packages/opencrvs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/language-opencrvs
 
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [213115b]
+- Updated dependencies [6ef5351]
+  - @openfn/language-fhir-4@0.2.10
+
 ## 1.0.16 - 24 February 2026
 
 ### Patch Changes